Made pmap code use int64_t where it used int (semi-dangerously) before.

+++ src/concurrency/pmap.hpp
@@ -10,9 +10,9 @@ template <class callable_t, class value_t>
 struct pmap_runner_one_arg_t {
     value_t i;
     const callable_t *c;
-    int *outstanding;
+    int64_t *outstanding;
     cond_t *to_signal;
-    pmap_runner_one_arg_t(value_t _i, const callable_t *_c, int *_outstanding,
+    pmap_runner_one_arg_t(value_t _i, const callable_t *_c, int64_t *_outstanding,
                           cond_t *_to_signal)
         : i(_i), c(_c), outstanding(_outstanding), to_signal(_to_signal) { }
 
@@ -26,12 +26,12 @@ struct pmap_runner_one_arg_t {
 };
 
 template <class callable_t, class value_t>
-void spawn_pmap_runner_one_arg(value_t i, const callable_t *c, int *outstanding, cond_t *to_signal) {
+void spawn_pmap_runner_one_arg(value_t i, const callable_t *c, int64_t *outstanding, cond_t *to_signal) {
     coro_t::spawn_now_dangerously(pmap_runner_one_arg_t<callable_t, value_t>(i, c, outstanding, to_signal));
 }
 
 template <class callable_t>
-void pmap(int begin, int end, const callable_t &c) {
+void pmap(int64_t begin, int64_t end, const callable_t &c) {
     guarantee(begin >= 0);  // We don't want `end - begin` to overflow, do we?
     guarantee(begin <= end);
     if (begin == end) {
@@ -39,24 +39,24 @@ void pmap(int begin, int end, const callable_t &c) {
     }
 
     cond_t cond;
-    int outstanding = (end - begin);
-    for (int i = begin; i < end - 1; ++i) {
-        coro_t::spawn_now_dangerously(pmap_runner_one_arg_t<callable_t, int>(i, &c, &outstanding, &cond));
+    int64_t outstanding = (end - begin);
+    for (int64_t i = begin; i < end - 1; ++i) {
+        coro_t::spawn_now_dangerously(pmap_runner_one_arg_t<callable_t, int64_t>(i, &c, &outstanding, &cond));
     }
-    pmap_runner_one_arg_t<callable_t, int> runner(end - 1, &c, &outstanding, &cond);
+    pmap_runner_one_arg_t<callable_t, int64_t> runner(end - 1, &c, &outstanding, &cond);
     runner();
     cond.wait();
 }
 
 template <class callable_t>
-void pmap(int count, const callable_t &c) {
+void pmap(int64_t count, const callable_t &c) {
     pmap(0, count, c);
 }
 
 template <class callable_t, class iterator_t>
 void pmap(iterator_t start, iterator_t end, const callable_t &c) {
     cond_t cond;
-    int outstanding = 1;
+    int64_t outstanding = 1;
     while (start != end) {
         ++outstanding;
         spawn_pmap_runner_one_arg(*start, &c, &outstanding, &cond);
@@ -72,10 +72,10 @@ template <class callable_t, class value_t>
 struct throttled_pmap_runner_t {
     value_t i;
     const callable_t *c;
-    int *outstanding;
+    int64_t *outstanding;
     cond_t *to_signal;
     new_semaphore_acq_t semaphore_acq;
-    throttled_pmap_runner_t(value_t _i, const callable_t *_c, int *_outstanding,
+    throttled_pmap_runner_t(value_t _i, const callable_t *_c, int64_t *_outstanding,
                             cond_t *_to_signal, new_semaphore_acq_t &&acq)
         : i(_i), c(_c), outstanding(_outstanding), to_signal(_to_signal),
           semaphore_acq(std::move(acq)) { }
@@ -90,9 +90,8 @@ struct throttled_pmap_runner_t {
     }
 };
 
-// TODO(2014-10): The types here are "int", but how do we know that won't overflow?
 template <class callable_t>
-void throttled_pmap(int begin, int end, const callable_t &c, int64_t capacity) {
+void throttled_pmap(int64_t begin, int64_t end, const callable_t &c, int64_t capacity) {
     guarantee(capacity > 0);
     guarantee(begin >= 0);  // We don't want `end - begin` to overflow, do we?
     guarantee(begin <= end);
@@ -102,19 +101,19 @@ void throttled_pmap(int begin, int end, const callable_t &c, int64_t capacity) {
 
     new_semaphore_t semaphore(capacity);
     cond_t cond;
-    int outstanding = (end - begin);
-    for (int i = begin; i < end - 1; ++i) {
+    int64_t outstanding = (end - begin);
+    for (int64_t i = begin; i < end - 1; ++i) {
         new_semaphore_acq_t acq(&semaphore, 1);
         acq.acquisition_signal()->wait();
         coro_t::spawn_now_dangerously(
-            throttled_pmap_runner_t<callable_t, int>(i, &c, &outstanding, &cond,
-                                                     std::move(acq)));
+            throttled_pmap_runner_t<callable_t, int64_t>(i, &c, &outstanding, &cond,
+                                                         std::move(acq)));
     }
 
     {
         new_semaphore_acq_t acq(&semaphore, 1);
         acq.acquisition_signal()->wait();
-        throttled_pmap_runner_t<callable_t, int> runner(end - 1, &c, &outstanding,
+        throttled_pmap_runner_t<callable_t, int64_t> runner(end - 1, &c, &outstanding,
                                                         &cond, std::move(acq));
         runner();
     }
@@ -122,7 +121,7 @@ void throttled_pmap(int begin, int end, const callable_t &c, int64_t capacity) {
 }
 
 template <class callable_t>
-void throttled_pmap(int count, const callable_t &c, int64_t capacity) {
+void throttled_pmap(int64_t count, const callable_t &c, int64_t capacity) {
     throttled_pmap(0, count, c, capacity);
 }
